% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dividends.R
\name{get_dividends}
\alias{get_dividends}
\title{Get stock market dividends via quantmod}
\usage{
get_dividends(
  symbols = "IBM",
  from = as.Date("1970-01-01"),
  to = as.Date("2022-12-31"),
  verbose = FALSE,
  warnings = TRUE,
  pad_by_month = FALSE
)
}
\arguments{
\item{symbols}{A character vector of stock market symbols to be accessed.}

\item{from}{A Date or string object that sets the starting date in YYYY-MM-DD ISO format.}

\item{to}{A Date or string object that sets the ending date in YYYY-MM-DD ISO format.}

\item{verbose}{A logical which if TRUE will output in verbose.}

\item{warnings}{A logical which if TRUE shows warnings.}

\item{pad_by_month}{A logical which if TRUE will pad the Date column with months having
dividend values of zero.}
}
\value{
A data table with historic "Date"'s along with dividend related variables.
}
\description{
A wrapper around quantmod::getDividends()) for returning a data table of historic dividends.
}
\author{
Rick Dean
}
