% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dow30_symbols.R
\name{get_dow30_symbols}
\alias{get_dow30_symbols}
\title{Get ticker symbols for "DOW 30" stocks that comprise the DOW
  Jones Industrial Average}
\usage{
get_dow30_symbols()
}
\value{
A data frame of ticker symbols and their company names.
}
\description{
Returns a data frame with character strings of ticker symbols
  for 30 of the stock market's top blue chip companies("ticker").
  Also a column that names the company ("company").
}
\author{
Rick Dean
}
