% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stock_prices.R
\name{get_stock_prices}
\alias{get_stock_prices}
\title{Get stock market prices via quantmod}
\usage{
get_stock_prices(
  symbols = "IBM",
  periodicity = "monthly",
  from = as.Date("2022-01-01"),
  to = as.Date("2022-12-31"),
  verbose = FALSE,
  source = "yahoo",
  output_csv = NULL,
  api_key = NULL,
  warnings = TRUE
)
}
\arguments{
\item{symbols}{A character vector of stock market symbols to be accessed.}

\item{periodicity}{A string that defines the frequency of sampled prices. Acceptable values are "daily",
"weekly", "monthly".}

\item{from}{A Date or string object that sets the starting date in YYYY-MM-DD ISO format. For \code{source}
equal to "tiingo", no sooner than "2007-01-01".}

\item{to}{A Date or string object that sets the ending date in YYYY-MM-DD ISO format.}

\item{verbose}{A logical which if TRUE will output in verbose.}

\item{source}{A string that sets the source for the data. Acceptable values are "yahoo" (default) or
"tiingo". If "tiingo" is selected then \code{periodicity} also has "annually".}

\item{output_csv}{A string that names a CSV file path to output the data.}

\item{api_key}{If "tiingo" is selected for \code{source}, then the api key issued by Tiingo is required.}

\item{warnings}{A logical which if TRUE shows warnings.}
}
\value{
A data table with historic "Date"'s along with stock price related variables.
}
\description{
A wrapper around quantmod::getSymbols()) for returning a data table of historic stock prices.
For more detailed info on getting prices from "yahoo" or "tiingo" see the help for
quantmod::getSymbols.yahoo() or quantmod::getSymbols.tiingo(). Note that stock price downloads
require an active internet connection.
}
\author{
Rick Dean
}
