% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tiingo_key.R
\name{read_tiingo_key}
\alias{read_tiingo_key}
\title{Read the tiingo authorization key for quantmod/riingo packages}
\usage{
read_tiingo_key()
}
\value{
returns the api key string
}
\description{
Function reads the \code{RIINGO_TOKEN} token from the .Renviron file.
}
\details{
A .Renviron file can be created by running \code{usethis::edit_r_environ()}.  A token
 named "RIINGO_TOKEN" was defined in this file with the user's tiingo auth key gotten from the tiingo web site.
 To make changes to .Renviron file enter \code{usethis::edit_r_environ()}.
}
